Add note on explicit method limits
Specifically, I note that the explicit numerical method is susceptible
to divergence from the analytical solution in this particular case
because the finite difference discretization can only be applied to
interior nodes. This divergence occurs when simulation time is too long
and/or the rod length considered is too short.