Add embedded gif
When the notebook is run, this change means that there are two animated
plots of the temperature distribution in the thin copper rod--one that
ca be controlled (Javascript), and another that loops repeatedly
(embedded gif). For users on Binder, the intent of this change is to
allow them to see the notebook's output without running the whole,
computationally intensive process.