Implement HDF5 data format, fix eqn (1) z step
Prior to this commit, the solution space was a preallocated numpy array
u[x, y, z, t] which had to be loaded into memory before a simulation
could run. By replacing this numpy array scheme with an HDF5 file which
is repeatedly written to and read from, required memory is replaced by
required disk space. This allows simulations to be run with more nodes
and finer time steps.

Additionally, A bug which involved future values computed during the z
step to be written along the x axis was corrected.