Fix overwriting results from solving in particular dimensions
In completing a finer grid solution, it was revealed that the ADI
procedure was overwriting solved future values in particular directions
because of the absence of partial time steps to which data could be
written. These partial time steps are now saved in the HDF5 file along
with the results of whole time steps. This required a new naming scheme
for time step data sets where "l_x_y" denotes the values at time step x
+ y/3.